+++ lib/pure/os.nim
@@ -987,7 +987,7 @@ when defined(windows) or defined(posix) or defined(nintendoswitch):
     ## When on Windows, it calls `quoteShellWindows proc
     ## <#quoteShellWindows,string>`_. Otherwise, calls `quoteShellPosix proc
     ## <#quoteShellPosix,string>`_.
-    ## See also: `system.quoted`.
+    ## See also: `strutils.quoted`.
     when defined(windows):
       return quoteShellWindows(s)
     else: